+++ job.go
@@ -57,7 +57,10 @@ func (r *JobService) List(ctx context.Context, query JobListParams, opts ...opti
 // After you delete a job, you can't recover it, but if you have services still
 // running with that job reference, they will re-create and re-schedule a new job
 // automatically.
-func (r *JobService) Delete(ctx context.Context, jobID int64, opts ...option.RequestOption) (res *string, err error) {
+func (r *JobService) Delete(ctx context.Context, jobID int64, body JobDeleteParams, opts ...option.RequestOption) (res *string, err error) {
+	if body.XAPIEnvironment.Present {
+		opts = append(opts, option.WithHeader("X-API-ENVIRONMENT", fmt.Sprintf("%s", body.XAPIEnvironment)))
+	}
 	opts = append(r.Options[:], opts...)
 	path := fmt.Sprintf("jobs/%v", jobID)
 	err = requestconfig.ExecuteNewRequest(ctx, http.MethodDelete, path, nil, &res, opts...)
@@ -72,9 +75,44 @@ func (r *JobService) Define(ctx context.Context, body JobDefineParams, opts ...o
 	return
 }
 
+// Temporary stops a job from running
+func (r *JobService) Pause(ctx context.Context, jobID int64, body JobPauseParams, opts ...option.RequestOption) (res *JobExecution, err error) {
+	if body.XAPIEnvironment.Present {
+		opts = append(opts, option.WithHeader("X-API-ENVIRONMENT", fmt.Sprintf("%s", body.XAPIEnvironment)))
+	}
+	opts = append(r.Options[:], opts...)
+	path := fmt.Sprintf("jobs/pause/%v", jobID)
+	err = requestconfig.ExecuteNewRequest(ctx, http.MethodPatch, path, nil, &res, opts...)
+	return
+}
+
+// Resumes job execution
+func (r *JobService) Resume(ctx context.Context, jobID int64, body JobResumeParams, opts ...option.RequestOption) (res *JobExecution, err error) {
+	if body.XAPIEnvironment.Present {
+		opts = append(opts, option.WithHeader("X-API-ENVIRONMENT", fmt.Sprintf("%s", body.XAPIEnvironment)))
+	}
+	opts = append(r.Options[:], opts...)
+	path := fmt.Sprintf("jobs/resume/%v", jobID)
+	err = requestconfig.ExecuteNewRequest(ctx, http.MethodPatch, path, nil, &res, opts...)
+	return
+}
+
+// Immediately triggers a job
+func (r *JobService) Trigger(ctx context.Context, jobID int64, body JobTriggerParams, opts ...option.RequestOption) (res *JobExecution, err error) {
+	if body.XAPIEnvironment.Present {
+		opts = append(opts, option.WithHeader("X-API-ENVIRONMENT", fmt.Sprintf("%s", body.XAPIEnvironment)))
+	}
+	opts = append(r.Options[:], opts...)
+	path := fmt.Sprintf("jobs/trigger/%v", jobID)
+	err = requestconfig.ExecuteNewRequest(ctx, http.MethodPost, path, nil, &res, opts...)
+	return
+}
+
 type Job struct {
 	// ID of the ent.
 	ID int64 `json:"id"`
+	// Blocking holds the value of the "blocking" field.
+	Blocking bool `json:"blocking"`
 	// Command to execute
 	Command string `json:"command"`
 	// Time when the job was created
@@ -95,12 +133,16 @@ type Job struct {
 	Name string `json:"name"`
 	// Scheduled time for next execution
 	NextRunAt string `json:"next_run_at"`
+	// Paused holds the value of the "paused" field.
+	Paused bool `json:"paused"`
 	// Number of retry attempts
 	RetryCount int64 `json:"retry_count"`
 	// Current job status (pending, running, completed, failed)
 	Status string `json:"status"`
 	// Maximum execution time before job is terminated
 	Timeout string `json:"timeout"`
+	// Maximum execution time before job is terminated
+	TimeoutSeconds int64 `json:"timeout_seconds"`
 	// Time when the job was last updated
 	UpdatedAt string  `json:"updated_at"`
 	JSON      jobJSON `json:"-"`
@@ -109,6 +151,7 @@ type Job struct {
 // jobJSON contains the JSON metadata for the struct [Job]
 type jobJSON struct {
 	ID             apijson.Field
+	Blocking       apijson.Field
 	Command        apijson.Field
 	CreatedAt      apijson.Field
 	CronExpression apijson.Field
@@ -119,9 +162,11 @@ type jobJSON struct {
 	Metadata       apijson.Field
 	Name           apijson.Field
 	NextRunAt      apijson.Field
+	Paused         apijson.Field
 	RetryCount     apijson.Field
 	Status         apijson.Field
 	Timeout        apijson.Field
+	TimeoutSeconds apijson.Field
 	UpdatedAt      apijson.Field
 	raw            string
 	ExtraFields    map[string]apijson.Field
@@ -165,14 +210,30 @@ type JobListParams struct {
 	XAPIEnvironment param.Field[int64] `header:"X-API-ENVIRONMENT,required"`
 }
 
+type JobDeleteParams struct {
+	XAPIEnvironment param.Field[int64] `header:"X-API-ENVIRONMENT,required"`
+}
+
 type JobDefineParams struct {
-	Name       param.Field[string]                 `json:"name,required"`
-	Schedule   param.Field[string]                 `json:"schedule,required"`
-	MaxRetries param.Field[int64]                  `json:"max_retries"`
-	Metadata   param.Field[map[string]interface{}] `json:"metadata"`
-	Timeout    param.Field[string]                 `json:"timeout"`
+	Name           param.Field[string]                 `json:"name,required"`
+	Schedule       param.Field[string]                 `json:"schedule,required"`
+	Blocking       param.Field[bool]                   `json:"blocking"`
+	Metadata       param.Field[map[string]interface{}] `json:"metadata"`
+	TimeoutSeconds param.Field[int64]                  `json:"timeout_seconds"`
 }
 
 func (r JobDefineParams) MarshalJSON() (data []byte, err error) {
 	return apijson.MarshalRoot(r)
+}
+
+type JobPauseParams struct {
+	XAPIEnvironment param.Field[int64] `header:"X-API-ENVIRONMENT,required"`
+}
+
+type JobResumeParams struct {
+	XAPIEnvironment param.Field[int64] `header:"X-API-ENVIRONMENT,required"`
+}
+
+type JobTriggerParams struct {
+	XAPIEnvironment param.Field[int64] `header:"X-API-ENVIRONMENT,required"`
 }