+++ job_test.go
@@ -80,7 +80,13 @@ func TestJobDelete(t *testing.T) {
 		option.WithBaseURL(baseURL),
 		option.WithAPIKey("My API Key"),
 	)
-	_, err := client.Jobs.Delete(context.TODO(), int64(0))
+	_, err := client.Jobs.Delete(
+		context.TODO(),
+		int64(0),
+		schedo.JobDeleteParams{
+			XAPIEnvironment: schedo.F(int64(0)),
+		},
+	)
 	if err != nil {
 		var apierr *schedo.Error
 		if errors.As(err, &apierr) {
@@ -104,14 +110,101 @@ func TestJobDefineWithOptionalParams(t *testing.T) {
 		option.WithAPIKey("My API Key"),
 	)
 	_, err := client.Jobs.Define(context.TODO(), schedo.JobDefineParams{
-		Name:       schedo.F("Name of your job"),
-		Schedule:   schedo.F("0 0 * * *"),
-		MaxRetries: schedo.F(int64(0)),
+		Name:     schedo.F("Name of your job"),
+		Schedule: schedo.F("0 0 * * *"),
+		Blocking: schedo.F(true),
 		Metadata: schedo.F(map[string]interface{}{
 			"foo": "bar",
 		}),
-		Timeout: schedo.F("timeout"),
+		TimeoutSeconds: schedo.F(int64(0)),
 	})
+	if err != nil {
+		var apierr *schedo.Error
+		if errors.As(err, &apierr) {
+			t.Log(string(apierr.DumpRequest(true)))
+		}
+		t.Fatalf("err should be nil: %s", err.Error())
+	}
+}
+
+func TestJobPause(t *testing.T) {
+	t.Skip("skipped: tests are disabled for the time being")
+	baseURL := "http://localhost:4010"
+	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
+		baseURL = envURL
+	}
+	if !testutil.CheckTestServer(t, baseURL) {
+		return
+	}
+	client := schedo.NewClient(
+		option.WithBaseURL(baseURL),
+		option.WithAPIKey("My API Key"),
+	)
+	_, err := client.Jobs.Pause(
+		context.TODO(),
+		int64(0),
+		schedo.JobPauseParams{
+			XAPIEnvironment: schedo.F(int64(0)),
+		},
+	)
+	if err != nil {
+		var apierr *schedo.Error
+		if errors.As(err, &apierr) {
+			t.Log(string(apierr.DumpRequest(true)))
+		}
+		t.Fatalf("err should be nil: %s", err.Error())
+	}
+}
+
+func TestJobResume(t *testing.T) {
+	t.Skip("skipped: tests are disabled for the time being")
+	baseURL := "http://localhost:4010"
+	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
+		baseURL = envURL
+	}
+	if !testutil.CheckTestServer(t, baseURL) {
+		return
+	}
+	client := schedo.NewClient(
+		option.WithBaseURL(baseURL),
+		option.WithAPIKey("My API Key"),
+	)
+	_, err := client.Jobs.Resume(
+		context.TODO(),
+		int64(0),
+		schedo.JobResumeParams{
+			XAPIEnvironment: schedo.F(int64(0)),
+		},
+	)
+	if err != nil {
+		var apierr *schedo.Error
+		if errors.As(err, &apierr) {
+			t.Log(string(apierr.DumpRequest(true)))
+		}
+		t.Fatalf("err should be nil: %s", err.Error())
+	}
+}
+
+func TestJobTrigger(t *testing.T) {
+	t.Skip("skipped: tests are disabled for the time being")
+	baseURL := "http://localhost:4010"
+	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
+		baseURL = envURL
+	}
+	if !testutil.CheckTestServer(t, baseURL) {
+		return
+	}
+	client := schedo.NewClient(
+		option.WithBaseURL(baseURL),
+		option.WithAPIKey("My API Key"),
+	)
+	_, err := client.Jobs.Trigger(
+		context.TODO(),
+		int64(0),
+		schedo.JobTriggerParams{
+			XAPIEnvironment: schedo.F(int64(0)),
+		},
+	)
 	if err != nil {
 		var apierr *schedo.Error
 		if errors.As(err, &apierr) {